Reorganize sync package
